Make QCoroTask header contain only declarations
Definitions are moved into private files inside the impl/ subdirectory.
This makes the header cleaner and easier to read and also solves repeated
problem with the need for out-of-line definitions for some methods as
they refer to other classes declared in the QCoroTask header.

This is a fully source-compatible change.
